Split out logging, plotting functionality from main
More progress on the quest to split things up.
